Move core types into Core folder and namespace.
