+++ source/Quantities.Benchmark/MeasureBenchmark.cs
@@ -1,4 +1,5 @@
 using BenchmarkDotNet.Attributes;
+using Quantities.Core;
 using Quantities.Measures;
 using Quantities.Prefixes;
 using Quantities.Units.Imperial.Length;
