+++ source/Quantities.Benchmark/Program.cs
@@ -1,4 +1,5 @@
-﻿using BenchmarkDotNet.Running;
+﻿global using Quantities.Core;
+using BenchmarkDotNet.Running;
 using Quantities.Benchmark;
 
 // dotnet run -c Release --project Quantities.Benchmark/
