+++ source/Quantities.Test/UserDefined.cs
@@ -1,5 +1,5 @@
 ﻿using Quantities.Dimensions;
-using Quantities.Numerics;
+using Quantities.Core.Numerics;
 using Quantities.Units;
 using Quantities.Units.NonStandard;
