+++ source/Quantities.Test/Usings.cs
@@ -1,4 +1,5 @@
-﻿global using Quantities.Prefixes;
+﻿global using Quantities.Core;
+global using Quantities.Prefixes;
 global using Quantities.Quantities;
 global using Quantities.Units.Imperial.Length;
 global using Quantities.Units.Si;
