+++ source/Quantities.Test/measures/ConversionTest.cs
@@ -1,5 +1,5 @@
 ﻿using Quantities.Measures;
-using Quantities.Numerics;
+using Quantities.Core.Numerics;
 using Quantities.Units.Imperial.Temperature;
 using Quantities.Units.Si.Derived;
 using Quantities.Units.Si.Metric;
