+++ source/Quantities.Test/measures/Expect.cs
@@ -1,5 +1,4 @@
-using Quantities.Measures;
-using Quantities.Numerics;
+using Quantities.Core.Numerics;
 
 namespace Quantities.Test.Measures;
