+++ source/Quantities.Test/prefixes/TestPrefix.cs
@@ -1,4 +1,4 @@
-﻿using Quantities.Numerics;
+﻿using Quantities.Core.Numerics;
 
 namespace Quantities.Test.Prefixes;
