+++ source/Quantities/Core/IMeasureEquality.cs
