+++ source/Quantities/Core/IRepresentable.cs
@@ -1,4 +1,4 @@
-﻿namespace Quantities;
+﻿namespace Quantities.Core;
 
 public interface IRepresentable
 {
