+++ source/Quantities/Core/ITransform.cs
@@ -1,4 +1,4 @@
-﻿namespace Quantities;
+﻿namespace Quantities.Core;
 
 public interface ITransform
 {
