+++ source/Quantities/Globals.cs
@@ -1,4 +1,5 @@
-﻿using System.Runtime.CompilerServices;
+﻿global using Quantities.Core;
+using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("quantities.test")]
 [assembly: InternalsVisibleTo("quantities.benchmark")]
